Create TaskUtil class for better separation
