+++ src/main/java/seedu/sherpass/task/TaskList.java
@@ -7,7 +7,6 @@ import seedu.sherpass.util.Ui;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.time.temporal.ChronoUnit;
 
 import java.util.ArrayList;
 import java.util.Comparator;
@@ -15,10 +14,6 @@ import java.util.HashSet;
 import java.util.Random;
 import java.util.stream.Collectors;
 
-import static seedu.sherpass.constant.Message.ERROR_BY_DATE_BEFORE_DO_ON_DATE;
-import static seedu.sherpass.constant.Message.ERROR_SCHEDULE_CLASH_MESSAGE;
-import static seedu.sherpass.constant.Message.ERROR_START_AFTER_END_TIME_MESSAGE;
-import static seedu.sherpass.constant.Message.ERROR_START_DATE_IN_THE_PAST_MESSAGE;
 import static seedu.sherpass.constant.Message.NEWLINE;
 import static seedu.sherpass.constant.Message.TAB_INDENT;
 
@@ -52,138 +47,23 @@ public class TaskList {
     }
 
     /**
-     * Returns the last recurrence of a recurring task.
-     * For tasks recurring daily, the last date is 1 month after the first occurrence.
-     * For tasks recurring weekly, the last date is 2 month after the first occurrence.
-     * For tasks recurring monthly, the last date is 1 year after the first occurrence.
-     *
-     * @param startDate The date of the first occurrence
-     * @param frequency The frequency of recurrence
-     * @return The date of the last recurrence for a recurring task.
-     */
-    private static LocalDateTime getEndDateForRecurrence(LocalDateTime startDate, Frequency frequency) {
-        if (frequency == Frequency.SINGLE) {
-            return startDate;
-        } else if (frequency == Frequency.DAILY) {
-            return startDate.plusMonths(1);
-        } else if (frequency == Frequency.WEEKLY) {
-            return startDate.plusMonths(2);
-        }
-        return startDate.plusYears(1);
-    }
-
-    private boolean isOnSameDay(LocalDateTime firstDate, LocalDateTime secondDate) {
-        return firstDate.toLocalDate().equals(secondDate.toLocalDate());
-    }
-
-    private boolean isEndTimeAfterCurrentStartTime(Task currentTask, LocalDateTime doOnStartDateTime,
-                                                   LocalDateTime doOnEndDateTime) {
-        return doOnEndDateTime.isAfter(currentTask.getDoOnStartDateTime())
-                && (doOnStartDateTime.isBefore(currentTask.getDoOnStartDateTime())
-                || doOnStartDateTime.equals(currentTask.getDoOnStartDateTime()));
-    }
-
-    private boolean isStartTimeBeforeCurrentEndTime(Task currentTask, LocalDateTime doOnStartDateTime,
-                                                    LocalDateTime doOnEndDateTime) {
-        return doOnStartDateTime.isBefore(currentTask.getDoOnEndDateTime())
-                && (doOnEndDateTime.isAfter(currentTask.getDoOnEndDateTime())
-                || doOnEndDateTime.equals(currentTask.getDoOnEndDateTime()));
-    }
-
-    private boolean isStartAndEndTimeWithinCurrentTime(Task currentTask, LocalDateTime doOnStartDateTime,
-                                                       LocalDateTime doOnEndDateTime) {
-        return doOnStartDateTime.isAfter(currentTask.getDoOnStartDateTime())
-                && doOnEndDateTime.isBefore(currentTask.getDoOnEndDateTime());
-    }
-
-    private boolean isStartAndEndTimeContainCurrentTime(Task currentTask, LocalDateTime doOnStartDateTime,
-                                                        LocalDateTime doOnEndDateTime) {
-        return currentTask.getDoOnStartDateTime().isAfter(doOnStartDateTime)
-                && currentTask.getDoOnEndDateTime().isBefore(doOnEndDateTime);
-    }
-
-    private boolean isStartAndEndTimeEqualsCurrentTime(Task currentTask, LocalDateTime doOnStartDateTime,
-                                                       LocalDateTime doOnEndDateTime) {
-        return doOnStartDateTime.isEqual(currentTask.getDoOnStartDateTime())
-                && doOnEndDateTime.equals(currentTask.getDoOnEndDateTime());
-    }
-
-    private boolean hasTimeClash(Task currentTask, LocalDateTime doOnStartDateTime,
-                                 LocalDateTime doOnEndDateTime) {
-        return isStartAndEndTimeEqualsCurrentTime(currentTask, doOnStartDateTime, doOnEndDateTime)
-                || isEndTimeAfterCurrentStartTime(currentTask, doOnStartDateTime, doOnEndDateTime)
-                || isStartTimeBeforeCurrentEndTime(currentTask, doOnStartDateTime, doOnEndDateTime)
-                || isStartAndEndTimeWithinCurrentTime(currentTask, doOnStartDateTime, doOnEndDateTime)
-                || isStartAndEndTimeContainCurrentTime(currentTask, doOnStartDateTime, doOnEndDateTime);
-    }
-
-    private boolean isStartTimeClashWithEndTime(Task taskToCheck) {
-        return taskToCheck.getDoOnStartDateTime().isAfter(taskToCheck.getDoOnEndDateTime())
-                || taskToCheck.getDoOnStartDateTime().equals(taskToCheck.getDoOnEndDateTime());
-    }
-
-    private boolean isByDateBeforeDoOnDate(Task taskToCheck) {
-        if (taskToCheck.getByDateTime() == null) {
-            return false;
-        }
-        return taskToCheck.getByDateTime().toLocalDate()
-                .isBefore(taskToCheck.getDoOnStartDateTime().toLocalDate());
-    }
-
-    /**
-     * Checks if there is any date and time clashes
-     * for a given array.
-     *
-     * @param taskList    Array representation of tasks.
-     * @param taskToCheck New Task to be checked for clash.
-     * @throws TimeClashException If there is a date and time clash, i.e.
-     *                            taskToCheck has the same date and clashing of time periods
-     *                            with tasks in taskList
-     */
-    public void checkDateTimeClash(ArrayList<Task> taskList, Task taskToCheck, boolean isFromFile)
-            throws TimeClashException, InvalidInputException {
-        if (isStartTimeClashWithEndTime(taskToCheck)) {
-            throw new InvalidInputException(ERROR_START_AFTER_END_TIME_MESSAGE);
-        }
-        if (!isFromFile && !taskToCheck.isDone() && taskToCheck.getDoOnStartDateTime().isBefore(LocalDateTime.now())) {
-            throw new InvalidInputException(ERROR_START_DATE_IN_THE_PAST_MESSAGE);
-        }
-        if (isByDateBeforeDoOnDate(taskToCheck)) {
-            throw new InvalidInputException(ERROR_BY_DATE_BEFORE_DO_ON_DATE);
-        }
-        for (Task task : taskList) {
-            if (isOnSameDay(task.getDoOnStartDateTime(), taskToCheck.getDoOnStartDateTime())
-                    && hasTimeClash(task, taskToCheck.getDoOnStartDateTime(), taskToCheck.getDoOnEndDateTime())) {
-                throw new TimeClashException(ERROR_SCHEDULE_CLASH_MESSAGE
-                        + "\n\t-> " + task.printTask() + "\n\t-> " + taskToCheck.printTask());
-            }
-        }
-    }
-
-    /**
      * Adds a new task to the current array of tasks.
      *
      * @param newTask The new task to be added to the array.
      */
     public void addTask(Task newTask, Frequency frequency, boolean isFromFile) throws InvalidInputException,
             TimeClashException {
-        LocalDateTime lastRecurrenceDate = getEndDateForRecurrence(newTask.getDoOnStartDateTime(),
+        LocalDateTime lastRecurrenceDate = TaskUtil.getEndDateForRecurrence(newTask.getDoOnStartDateTime(),
                 frequency);
         ArrayList<Task> taskListToAdd = new ArrayList<>();
+        Task currentTask = newTask;
         do {
-            checkDateTimeClash(tasks, newTask, isFromFile);
-            taskListToAdd.add(newTask);
-            newTask = newTask.prepareNextTask(frequency);
-        } while (newTask.getDoOnStartDateTime().isBefore(lastRecurrenceDate));
+            TaskUtil.checkDateTimeClash(tasks, currentTask, isFromFile);
+            taskListToAdd.add(currentTask);
+            currentTask = TaskUtil.prepareNextTask(currentTask, frequency);
+        } while (currentTask.getDoOnStartDateTime().isBefore(lastRecurrenceDate));
         tasks.addAll(taskListToAdd);
         updateIndex();
-    }
-
-    private long calculateOffsetOfDate(LocalDateTime oldDateTime, LocalDateTime newDateTime) {
-        if (oldDateTime != null && newDateTime != null) {
-            return oldDateTime.until(newDateTime, ChronoUnit.SECONDS);
-        }
-        return 0;
     }
 
     /**
@@ -198,24 +78,24 @@ public class TaskList {
      * @throws TimeClashException    if the new do date has is clashing with other tasks
      * @throws InvalidInputException if the dates are invalid e.g. start date after end date
      */
-    public Task editSingleTaskContent(int editIndex, String taskDescription,
-                                      LocalDateTime doOnStartDateTime,
-                                      LocalDateTime doOnEndDateTime,
-                                      LocalDateTime byDateTime) throws TimeClashException, InvalidInputException {
+    public Task editSingleTask(int editIndex, String taskDescription,
+                               LocalDateTime doOnStartDateTime,
+                               LocalDateTime doOnEndDateTime,
+                               LocalDateTime byDateTime) throws TimeClashException, InvalidInputException {
 
         Task taskToEdit = tasks.get(editIndex);
-        ArrayList<Task> editedList = new ArrayList<>(tasks);
-        editedList.remove(editIndex);
+        ArrayList<Task> tempList = new ArrayList<>(tasks);
+        tempList.remove(editIndex);
 
-        long startDateOffset = calculateOffsetOfDate(taskToEdit.getDoOnStartDateTime(), doOnStartDateTime);
-        long endDateOffset = calculateOffsetOfDate(taskToEdit.getDoOnEndDateTime(), doOnEndDateTime);
-        long byDateOffset = calculateOffsetOfDate(doOnStartDateTime, byDateTime);
+        long startDateOffset = TaskUtil.calculateOffsetOfDate(taskToEdit.getDoOnStartDateTime(), doOnStartDateTime);
+        long endDateOffset = TaskUtil.calculateOffsetOfDate(taskToEdit.getDoOnEndDateTime(), doOnEndDateTime);
+        long byDateOffset = TaskUtil.calculateOffsetOfDate(doOnStartDateTime, byDateTime);
 
         Task updatedTask = taskToEdit.clone();
         updatedTask.editTask(generateIdentifier(), taskDescription,
                 startDateOffset, endDateOffset, byDateOffset);
 
-        checkDateTimeClash(editedList, updatedTask, false);
+        TaskUtil.checkDateTimeClash(tempList, updatedTask, false);
 
         tasks.remove(editIndex);
         tasks.add(updatedTask);
@@ -241,14 +121,14 @@ public class TaskList {
                                   LocalDateTime byDateTime) throws TimeClashException, InvalidInputException {
         Task firstTask = getTask(editIndex);
         ArrayList<Task> affectedTasks = getAffectedTasks(editIndex);
-        ArrayList<Task> editedList = new ArrayList<>(tasks);
+        ArrayList<Task> tempList = new ArrayList<>(tasks);
         ArrayList<Task> editedTasks = new ArrayList<>();
         assert (affectedTasks.size() > 0);
-        editedList.removeAll(affectedTasks);
+        tempList.removeAll(affectedTasks);
 
-        long startDateOffset = calculateOffsetOfDate(firstTask.getDoOnStartDateTime(), doOnStartDateTime);
-        long endDateOffset = calculateOffsetOfDate(firstTask.getDoOnEndDateTime(), doOnEndDateTime);
-        long byDateOffset = calculateOffsetOfDate(doOnStartDateTime, byDateTime);
+        long startDateOffset = TaskUtil.calculateOffsetOfDate(firstTask.getDoOnStartDateTime(), doOnStartDateTime);
+        long endDateOffset = TaskUtil.calculateOffsetOfDate(firstTask.getDoOnEndDateTime(), doOnEndDateTime);
+        long byDateOffset = TaskUtil.calculateOffsetOfDate(doOnStartDateTime, byDateTime);
 
         int newIdentifier = generateIdentifier();
         for (Task t : affectedTasks) {
@@ -256,12 +136,12 @@ public class TaskList {
             updatedTask.editTask(newIdentifier, taskDescription,
                     startDateOffset, endDateOffset, byDateOffset);
             updatedTask.setIdentifier(newIdentifier);
-            checkDateTimeClash(editedList, updatedTask, false);
+            TaskUtil.checkDateTimeClash(tempList, updatedTask, false);
             editedTasks.add(updatedTask);
-            editedList.add(updatedTask);
+            tempList.add(updatedTask);
         }
 
-        tasks = editedList;
+        tasks = tempList;
         updateIndex();
         return editedTasks.get(0);
     }
