+++ src/test/java/seedu/sherpass/task/TaskListTest.java
@@ -24,9 +24,8 @@ public class TaskListTest {
         Task testTaskTwo = new Task(2, "smt", null,
                 LocalDateTime.parse("31/3/2023 10:00", inputWithTimeFormat),
                 LocalDateTime.parse("31/3/2023 13:00", inputWithTimeFormat));
-        TaskList testMethod = new TaskList();
         assertThrows(TimeClashException.class,
-            () -> testMethod.checkDateTimeClash(dummyList, testTaskTwo, false));
+            () -> TaskUtil.checkDateTimeClash(dummyList, testTaskTwo, false));
     }
 
     @Test
@@ -39,9 +38,8 @@ public class TaskListTest {
         Task testTaskTwo = new Task(2, "smt", null,
                 LocalDateTime.parse("31/3/2023 11:00", inputWithTimeFormat),
                 LocalDateTime.parse("31/3/2023 12:00", inputWithTimeFormat));
-        TaskList testMethod = new TaskList();
         assertThrows(TimeClashException.class,
-            () -> testMethod.checkDateTimeClash(dummyList, testTaskTwo, false));
+            () -> TaskUtil.checkDateTimeClash(dummyList, testTaskTwo, false));
     }
 
     @Test
@@ -54,9 +52,8 @@ public class TaskListTest {
         Task testTaskTwo = new Task(2, "smt", null,
                 LocalDateTime.parse("31/3/2023 10:30", inputWithTimeFormat),
                 LocalDateTime.parse("31/3/2023 11:00", inputWithTimeFormat));
-        TaskList testMethod = new TaskList();
         assertThrows(TimeClashException.class,
-            () -> testMethod.checkDateTimeClash(dummyList, testTaskTwo, false));
+            () -> TaskUtil.checkDateTimeClash(dummyList, testTaskTwo, false));
     }
 
     @Test
@@ -69,9 +66,8 @@ public class TaskListTest {
         Task testTaskTwo = new Task(2, "smt", null,
                 LocalDateTime.parse("31/3/2023 09:00", inputWithTimeFormat),
                 LocalDateTime.parse("31/3/2023 13:00", inputWithTimeFormat));
-        TaskList testMethod = new TaskList();
         assertThrows(TimeClashException.class,
-            () -> testMethod.checkDateTimeClash(dummyList, testTaskTwo, false));
+            () -> TaskUtil.checkDateTimeClash(dummyList, testTaskTwo, false));
     }
 
     @Test
@@ -84,8 +80,7 @@ public class TaskListTest {
         Task testTaskTwo = new Task(2, "break time!", null,
                 LocalDateTime.parse("31/3/2023 12:00", inputWithTimeFormat),
                 LocalDateTime.parse("31/3/2023 13:00", inputWithTimeFormat));
-        TaskList testMethod = new TaskList();
-        testMethod.checkDateTimeClash(dummyList, testTaskTwo, false);
+        TaskUtil.checkDateTimeClash(dummyList, testTaskTwo, false);
         assertTrue(true);
     }
